+++ src/components/Canvas/StickerPicker.tsx
