+++ src/components/CanvasLayer/Note.tsx
