+++ src/components/CanvasLayer/Sticker.tsx
