+++ src/components/Layout/CollabToolAside/MusicPlayer/WaveForm.tsx
